refactor(background): name retained handlers (#284)
